+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 [Full diff](https://github.com/sider/runners/compare/0.21.0...HEAD)
 
 - [GolangCI-Lint] Warn no Go files instead of failing [#815](https://github.com/sider/runners/pull/815)
+- Improve trace message on start and finish [#816](https://github.com/sider/runners/pull/816)
 
 ## 0.21.0